test(AdminControllerTest): refactoring
