+++ main/src/test/kotlin/com/ip13/main/controller/AdminControllerTest.kt
@@ -1,15 +1,23 @@
 package com.ip13.main.controller
 
 import com.ip13.main.security.repository.UserRepository
+import com.ip13.main.utils.loadAsString
+import org.hamcrest.CoreMatchers
 import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc
 import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.http.MediaType
 import org.springframework.jdbc.core.JdbcTemplate
 import org.springframework.security.crypto.password.PasswordEncoder
+import org.springframework.security.test.context.support.WithMockUser
 import org.springframework.test.context.ActiveProfiles
 import org.springframework.test.context.DynamicPropertyRegistry
 import org.springframework.test.context.DynamicPropertySource
+import org.springframework.test.web.servlet.MockMvc
+import org.springframework.test.web.servlet.post
+import org.springframework.test.web.servlet.result.MockMvcResultMatchers
 import org.testcontainers.containers.PostgreSQLContainer
 import org.testcontainers.junit.jupiter.Container
 import org.testcontainers.junit.jupiter.Testcontainers
@@ -26,6 +34,10 @@ class AdminControllerTest(
     @Autowired val userRepository: UserRepository,
     @Autowired val passwordEncoder: PasswordEncoder,
 ) {
+    @Autowired
+    private lateinit var mockMvc: MockMvc
+
+
     @BeforeEach
     fun cleanUp() {
         jdbc.execute("truncate table black_list cascade")
@@ -36,6 +48,28 @@ class AdminControllerTest(
         jdbc.execute("truncate table restaurant cascade")
         jdbc.execute("truncate table restaurant_add_ticket cascade")
         jdbc.execute("delete from user_t where id != 100")
+    }
+
+    @Test
+    @WithMockUser(authorities = [ADMIN])
+    fun `test add role to non-existent user`() {
+        val body = loadAsString("json/non_existent_user.json")
+
+        mockMvc.post("/admin/add_role") {
+            contentType = MediaType.APPLICATION_JSON
+            accept = MediaType.APPLICATION_JSON
+            content = body
+        }.andExpect {
+            MockMvcResultMatchers.status().`is`(400)
+            content {
+                jsonPath(
+                    "message",
+                    CoreMatchers.containsString(
+                        "UserNotFoundException: User with id: 10 not found",
+                    )
+                )
+            }
+        }
     }
 
     companion object {
